koutoubi-backend/requirements.txt
# Core
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
python-dotenv==1.0.0

# Database
sqlalchemy==2.0.25
alembic==1.13.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
supabase==2.3.4

# Validation
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# PDF Processing
PyPDF2==3.0.1
pypdfium2==4.26.0
PyMuPDF==1.23.19
pdfplumber==0.10.3
pycryptodome==3.23.0

# AI/LLM
langchain==0.1.1
langchain-community==0.0.12
openai==1.9.0
tiktoken==0.5.2
ollama==0.1.6

# Vector Store
chromadb==0.4.22
sentence-transformers==2.2.2

# Content Processing
markdown==3.5.2
beautifulsoup4==4.12.3
lxml==5.1.0

# Cache
redis==5.0.1
aiocache==0.12.2

# Utils
httpx>=0.26.0
aiofiles==23.2.1
python-dateutil==2.8.2
pytz==2023.3

# Security & Rate Limiting
slowapi==0.1.9
cryptography==41.0.7

# CORS
fastapi-cors==0.0.6

# Development
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
black==23.12.1
flake8==7.0.0
mypy==1.8.0
pre-commit==3.6.0

# Documentation
mkdocs==1.5.3
mkdocs-material==9.5.3

# Monitoring
prometheus-client==0.19.0
loguru==0.7.2